Expand test suite to 178 tests and fix additional bugs
- Add pytest as dev dependency in pyproject.toml
- Fix sawtooth_wave: numpy.linspace requires int for num parameter
- Fix Scale.degree() lookup by name ("tonic", "dominant") was broken
- Add tests for: Tone.from_tuple, from_index, _index, error paths,
  multi-octave arithmetic, descending chromatic walk, all pitch frequencies,
  Pythagorean temperament, all scale degrees by Roman numeral and name,
  Scale repr/slice/degrees mismatch, TonedScale repr/get/with-Tone-object,
  D/F/A/E/B major/minor scales, D Dorian, G Mixolydian, octave boundary
  crossing, all mode intervals including Phrygian/Aeolian/Locrian/Ionian,
  chord intervals/beat_pulse/empty chord, Chord/Fretboard fingering errors,
  all NamedChord qualities (m6/m9/maj9/9), fix_fingering, flat-to-sharp
  conversion, charts total count, charts_for_fretboard, all wave generators,
  wave output types, Synth with real pitches, circle of fifths/fourths,
  relative minor equivalence, enharmonic equivalence in scales, I-IV-V
  progression integration test

+++ pytheory/play.py
@@ -26,7 +26,7 @@ def sawtooth_wave(hz, peak=SAMPLE_PEAK, rising_ramp_width=1, n_samples=SAMPLE_RA
     rising_ramp_width is the percentage of the ramp spend rising:
     .5 is a triangle wave with equal rising and falling times.
     """
-    t = numpy.linspace(0, 1, 500 * 440 / hz, endpoint=False)
+    t = numpy.linspace(0, 1, int(500 * 440 / hz), endpoint=False)
     wave = scipy.signal.sawtooth(2 * numpy.pi * 5 * t, width=rising_ramp_width)
     wave = numpy.resize(wave, (n_samples,))
     # Sawtooth waves sound very quiet, so multiply peak by 4.
